Move the `OpenTypeFileBuilder` from `src/core/fonts.js` and into its own file
